Changed the interactors from github to their respective files.
