+++ app/interactors/repositories.js
@@ -5,12 +5,7 @@ const {
   addCodeownersToRepo: addCodeownersToRepoGithub,
   getRepositories: getRepositoriesGithub
 } = require('../services/github/repositories');
-const {
-  getTeams: getTeamsGithub,
-  createTeam: createTeamGithub,
-  addMemberToTeam: addMemberToTeamGithub,
-  deleteTeam: deleteTeamGithub
-} = require('../services/github/teams');
+
 const {
   createBranchFromMaster,
   updateBranchProtection,
@@ -18,7 +13,6 @@ const {
   DEVELOPMENT_BRANCH,
   STAGE_BRANCH
 } = require('../services/github/branches');
-const { addUser } = require('../services/github/organization');
 
 const execDefaultRepositoryActions = ({ repositoryName }) =>
   Promise.all([
@@ -38,30 +32,15 @@ const createRepository = ({ repositoryName, isPrivate }) =>
     execDefaultRepositoryActions({ repositoryName }).then(() => repository)
   );
 
-const getTeams = (page, limit) =>
-  getTeamsGithub(page, limit).then(resp => ({ teams: resp.data.map(({ name, id }) => ({ name, id })) }));
-
-const createTeam = name => createTeamGithub(name).then(resp => resp.data);
-
 const addTeamToRepo = (teamId, repositoryName) =>
   addTeamToRepository({ teamId, repositoryName }).then(resp => resp.data);
-
-const addMemberToTeam = (teamId, username) =>
-  addMemberToTeamGithub({ teamId, username }).then(resp => resp.data);
 
 const addCodeownersToRepo = (repositoryName, codeowners) =>
   addCodeownersToRepoGithub({ repositoryName, codeowners }).then(resp => resp.data);
 
-const deleteTeam = teamId => deleteTeamGithub({ teamId }).then(resp => resp.data);
-
 module.exports = {
   createRepository,
   getRepositories,
-  getTeams,
-  createTeam,
   addTeamToRepo,
-  addMemberToTeam,
-  addCodeownersToRepo,
-  deleteTeam,
-  addUser
+  addCodeownersToRepo
 };
